// src/app/shared/enums/card-payment-enums.ts
export enum MessageType {
    REQ_WS_INIT = "REQ_WS_INIT",
    REQ_WS_EXIT = "REQ_WS_EXIT",
    REQ_WS_RESTART = "REQ_WS_RESTART",
    REQ_INIT_WITH_CONFIGURATION = "REQ_INIT_WITH_CONFIGURATION",
    REQ_INIT_DEVICE = "REQ_INIT_DEVICE",
    REQ_DISCONNECT_DEVICE = "REQ_DISCONNECT_DEVICE",
    REQ_SET_LOG_LEVEL = "REQ_SET_LOG_LEVEL",
    REQ_SET_MODE = "REQ_SET_MODE",
    REQ_PROCESS_SALE = "REQ_PROCESS_SALE",
    REQ_PROCESS_REFUND = "REQ_PROCESS_REFUND",
    REQ_PROCESS_UNREFERENCED_REFUND = "REQ_PROCESS_UNREFERENCED_REFUND",
    REQ_SUBMIT_SIGNATURE = "REQ_SUBMIT_SIGNATURE",
    REQ_RETRIEVE_SETTINGS = "REQ_RETRIEVE_SETTINGS",
    REQ_LIST_SUPPORTED_DEVICES = "REQ_LIST_SUPPORTED_DEVICES",
    REQ_SET_DELAYED_AUTH_ENABLED = "REQ_SET_DELAYED_AUTH_ENABLED",
    REQ_SET_POLLING_ENABLED = "REQ_SET_POLLING_ENABLED",
    REQ_GET_DELAYED_AUTH_ENABLED = "REQ_GET_DELAYED_AUTH_ENABLED",
    REQ_GET_POLLING_ENABLED = "REQ_GET_POLLING_ENABLED",
    REQ_SET_TERMINAL_HEALTH_LISTENER_ENABLED = "REQ_SET_TERMINAL_HEALTH_LISTENER_ENABLED",
    REQ_SET_OFFLINE_MODE_ENABLED = "REQ_SET_OFFLINE_MODE_ENABLED",
    REQ_CANCEL_TRANSACTION = "REQ_CANCEL_TRANSACTION",
    REQ_GET_TRANSACTIONS = "REQ_GET_TRANSACTIONS",
    REQ_CLOSE_TRANSACTIONS_BATCH = "REQ_CLOSE_TRANSACTIONS_BATCH",
    REQ_GET_CALCULATED_AMOUNT = "REQ_GET_CALCULATED_AMOUNT",
    REQ_SUBMIT_APPLICATION = "REQ_SUBMIT_APPLICATION",
    REQ_SUBMIT_AMOUNT = "REQ_SUBMIT_AMOUNT",
    REQ_GET_DEVICE_INFO = "REQ_GET_DEVICE_INFO",
    RES_ON_WS_ERROR = "RES_ON_WS_ERROR",
    RES_ON_SETTINGS_RETRIEVED = "RES_ON_SETTINGS_RETRIEVED",
    RES_ON_ERROR = "RES_ON_ERROR",
    RES_ON_DEVICE_ERROR = "RES_ON_DEVICE_ERROR",
    RES_ON_SALE_RESPONSE = "RES_ON_SALE_RESPONSE",
    RES_ON_SIGNATURE_REQUIRED = "RES_ON_SIGNATURE_REQUIRED",
    RES_ON_DEVICE_CONNECTED = "RES_ON_DEVICE_CONNECTED",
    RES_ON_DEVICE_DISCONNECTED = "RES_ON_DEVICE_DISCONNECTED",
    RES_ON_MESSAGE = "RES_ON_MESSAGE",
    RES_ON_SUPPORTED_DEVICES_RETRIEVED = "RES_ON_SUPPORTED_DEVICES_RETRIEVED",
    RES_ON_WS_INIT_REQUIRED = "RES_ON_WS_INIT_REQUIRED",
    RES_ON_WS_INIT = "RES_ON_WS_INIT",
    RES_ON_WS_EXIT = "RES_ON_WS_EXIT",
    RES_ON_WS_RESTART = "RES_ON_WS_RESTART",
    RES_ON_SELECT_APPLICATION = "RES_ON_SELECT_APPLICATION",
    RES_ON_DEVICE_CONNECTION_ERROR = "RES_ON_DEVICE_CONNECTION_ERROR",
    RES_ON_AUTO_CONFIG_PROGRESS_UPDATE = "RES_ON_AUTO_CONFIG_PROGRESS_UPDATE",
    RES_ON_SELECT_SERIAL_PORT = "RES_ON_SELECT_SERIAL_PORT",
    RES_ON_DEVICE_INFO_RETURNED = "RES_ON_DEVICE_INFO_RETURNED",
    RES_ON_REFUND_RESPONSE = "RES_ON_REFUND_RESPONSE",
    RES_ON_REVERSAL_RETRIEVED = "RES_ON_REVERSAL_RETRIEVED",
    RES_ON_OFFLINE_SALE_REQUEST = "RES_ON_OFFLINE_SALE_REQUEST",
    RES_ON_REQUEST_SET_AMOUNT = "RES_ON_REQUEST_SET_AMOUNT",
    RES_ON_GIFT_CARD_DATA_RETURNED = "RES_ON_GIFT_CARD_DATA_RETURNED",
    RES_ON_REQUEST_UPDATE_RESPONSE = "RES_ON_REQUEST_UPDATE_RESPONSE",
    RES_ON_LOG_MESSAGE = "RES_ON_LOG_MESSAGE",
    RES_ON_TRANSACTION_LIST_RESPONSE = "RES_ON_TRANSACTION_LIST_RESPONSE",
    RES_ON_REQUEST_CLOSE_BATCH_RESPONSE = "RES_ON_REQUEST_CLOSE_BATCH_RESPONSE",
    RES_ON_LOGIN_URL_RETRIEVED = "RES_ON_LOGIN_URL_RETRIEVED",
    RES_ON_SECURE_CARD_RESPONSE = "RES_ON_SECURE_CARD_RESPONSE",
    RES_ON_SET_LOG_LEVEL = "RES_ON_SET_LOG_LEVEL",
    RES_ON_SET_MODE = "RES_ON_SET_MODE",
    RES_ON_SET_DELAYED_AUTH_ENABLED = "RES_ON_SET_DELAYED_AUTH_ENABLED",
    RES_ON_SET_POLLING_ENABLED = "RES_ON_SET_POLLING_ENABLED",
    RES_ON_GET_DELAYED_AUTH_ENABLED = "RES_ON_GET_DELAYED_AUTH_ENABLED",
    RES_ON_GET_POLLING_ENABLED = "RES_ON_GET_POLLING_ENABLED",
    RES_ON_SET_TERMINAL_HEALTH_LISTENER_ENABLED = "RES_ON_SET_TERMINAL_HEALTH_LISTENER_ENABLED",
    RES_ON_TERMINAL_HEALTH_LISTENER_STARTED = "RES_ON_TERMINAL_HEALTH_LISTENER_STARTED",
    RES_ON_TERMINAL_HEALTH_LISTENER_STOPPED = "RES_ON_TERMINAL_HEALTH_LISTENER_STOPPED",
    RES_ON_TERMINAL_HEALTH_LISTENER_ERROR = "RES_ON_TERMINAL_HEALTH_LISTENER_ERROR",
    RES_ON_SET_OFFLINE_MODE_ENABLED = "RES_ON_SET_OFFLINE_MODE_ENABLED",
    RES_ON_CALCULATED_AMOUNT_RESPONSE = "RES_ON_CALCULATED_AMOUNT_RESPONSE",

}